Apply `backend.result_type` to `count_nonzero`

+++ keras/backend/tensorflow/numpy.py
@@ -542,7 +542,7 @@ def cosh(x):
 
 
 def count_nonzero(x, axis=None):
-    return tfnp.count_nonzero(x, axis=axis)
+    return tf.cast(tfnp.count_nonzero(x, axis=axis), "int32")
 
 
 def cross(x1, x2, axisa=-1, axisb=-1, axisc=-1, axis=None):
